refactor: cleaned up azure service names and unused methods
