+++ src/Sepes.Functions/Service/FunctionUserService.cs
@@ -10,7 +10,7 @@ namespace Sepes.Infrastructure.Service
         
         public FunctionUserService()
         {
-            _cachedUser = new UserDto("9b0c65cf-9f14-4476-8796-b2de016e1af1", "workeruser@equinor.com", "Worker User", "workeruser@equinor.com", false, false, false);
+            _cachedUser = new UserDto("9b0c65cf-9f14-4476-8796-b2de016e1af1", "workeruser@equinor.com", "Worker User", "workeruser@equinor.com", admin: true, false, false);
         }
 
 #pragma warning disable CS1998 // Async method lacks 'await' operators and will run synchronously
