+++ src/Sepes.Infrastructure/Service/Azure/Interface/IAzureRoleAssignmentService.cs
@@ -1,5 +1,4 @@
 ﻿using Sepes.Infrastructure.Dto;
-using Sepes.Infrastructure.Dto.Azure.RoleAssignment;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
@@ -8,11 +7,11 @@ namespace Sepes.Infrastructure.Service.Azure.Interface
 {
     public interface IAzureRoleAssignmentService
     {
-        Task<bool> RoleAssignmentExists(string resourceId, string roleAssignmentId, CancellationToken cancellationToken = default);
+        //Task<bool> RoleAssignmentExists(string resourceId, string roleAssignmentId, CancellationToken cancellationToken = default);
 
-        Task<AzureRoleAssignment> DeleteRoleAssignment(string roleAssignmentId, CancellationToken cancellationToken = default);
+        //Task<AzureRoleAssignment> DeleteRoleAssignment(string roleAssignmentId, CancellationToken cancellationToken = default);
        
-        Task<AzureRoleAssignment> AddRoleAssignment(string resourceId, string roleDefinitionId, string principalId, string roleAssignmentId = null, CancellationToken cancellationToken = default);
+        //Task<AzureRoleAssignment> AddRoleAssignment(string resourceId, string roleDefinitionId, string principalId, string roleAssignmentId = null, CancellationToken cancellationToken = default);
 
         Task SetRoleAssignments(string resourceGroupId, string resourceGroupName, List<CloudResourceDesiredRoleAssignmentDto> desiredRoleAssignments, CancellationToken cancellationToken = default);
