+++ src/Sepes.Infrastructure/Service/Azure/Interface/IAzureVirtualMachineExtenedInfoService.cs
@@ -1,0 +1,12 @@
+﻿using Sepes.Infrastructure.Dto.VirtualMachine;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Sepes.Infrastructure.Service.Azure.Interface
+{
+    public interface IAzureVirtualMachineExtenedInfoService
+    {
+        Task<VmExtendedDto> GetExtendedInfo(string resourceGroupName, string resourceName, CancellationToken cancellationToken = default);
+    }
+}
+
