+++ src/Sepes.Infrastructure/Service/Azure/Interface/IAzureVirtualMachineService.cs
@@ -1,0 +1,9 @@
+﻿namespace Sepes.Infrastructure.Service.Azure.Interface
+{
+    public interface IAzureVirtualMachineService : IHasProvisioningState, IHasTags, IPerformResourceProvisioning
+    {     
+
+        
+    }
+}
+
