+++ src/Sepes.Infrastructure/Service/Azure/Interface/IAzureVirtualNetworkService.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Sepes.Infrastructure.Service.Azure.Interface
+{
+    public interface IAzureVirtualNetworkService : IHasProvisioningState, IHasTags, IPerformResourceProvisioning
+    {       
+        Task EnsureSandboxSubnetHasServiceEndpointForStorage(string resourceGroupName, string networkName);
+    }
+}
