+++ src/Sepes.Infrastructure/Util/Azure/AzureResourceServiceResolver.cs
@@ -14,22 +14,22 @@ namespace Sepes.Infrastructure.Util
         public static IHasProvisioningState GetServiceWithProvisioningState(IServiceProvider serviceProvider, string resourceType) => resourceType switch
         {
             AzureResourceType.ResourceGroup => serviceProvider.GetRequiredService<IAzureResourceGroupService>(),
-            AzureResourceType.VirtualNetwork => serviceProvider.GetRequiredService<IAzureVNetService>(),
+            AzureResourceType.VirtualNetwork => serviceProvider.GetRequiredService<IAzureVirtualNetworkService>(),
             AzureResourceType.NetworkSecurityGroup => serviceProvider.GetRequiredService<IAzureNetworkSecurityGroupService>(),
             AzureResourceType.StorageAccount => serviceProvider.GetRequiredService<IAzureStorageAccountService>(),
             AzureResourceType.Bastion => serviceProvider.GetRequiredService<IAzureBastionService>(),
-            AzureResourceType.VirtualMachine => serviceProvider.GetRequiredService<IAzureVmService>(),
+            AzureResourceType.VirtualMachine => serviceProvider.GetRequiredService<IAzureVirtualMachineService>(),
             _ => null,
         };
 
         public static IHasTags GetServiceWithTags(IServiceProvider serviceProvider, string resourceType) => resourceType switch
         {
             AzureResourceType.ResourceGroup => serviceProvider.GetRequiredService<IAzureResourceGroupService>(),
-            AzureResourceType.VirtualNetwork => serviceProvider.GetRequiredService<IAzureVNetService>(),
+            AzureResourceType.VirtualNetwork => serviceProvider.GetRequiredService<IAzureVirtualNetworkService>(),
             AzureResourceType.NetworkSecurityGroup => serviceProvider.GetRequiredService<IAzureNetworkSecurityGroupService>(),
             AzureResourceType.StorageAccount => serviceProvider.GetRequiredService<IAzureStorageAccountService>(),
             AzureResourceType.Bastion => serviceProvider.GetRequiredService<IAzureBastionService>(),
-            AzureResourceType.VirtualMachine => serviceProvider.GetRequiredService<IAzureVmService>(),
+            AzureResourceType.VirtualMachine => serviceProvider.GetRequiredService<IAzureVirtualMachineService>(),
             _ => null,
         };
 
@@ -38,9 +38,9 @@ namespace Sepes.Infrastructure.Util
             AzureResourceType.ResourceGroup => serviceProvider.GetRequiredService<IAzureResourceGroupService>(),
             AzureResourceType.NetworkSecurityGroup => serviceProvider.GetRequiredService<IAzureNetworkSecurityGroupService>(),
             AzureResourceType.StorageAccount => serviceProvider.GetRequiredService<IAzureStorageAccountService>(),
-            AzureResourceType.VirtualNetwork => serviceProvider.GetRequiredService<IAzureVNetService>(),
+            AzureResourceType.VirtualNetwork => serviceProvider.GetRequiredService<IAzureVirtualNetworkService>(),
             AzureResourceType.Bastion => serviceProvider.GetRequiredService<IAzureBastionService>(),
-            AzureResourceType.VirtualMachine => serviceProvider.GetRequiredService<IAzureVmService>(),
+            AzureResourceType.VirtualMachine => serviceProvider.GetRequiredService<IAzureVirtualMachineService>(),
             _ => null,
         };
