+++ src/Sepes.Infrastructure/Util/Provisioning/EnsureCorsRulesUtil.cs
@@ -38,7 +38,7 @@ namespace Sepes.Infrastructure.Util.Provisioning
 
                 if (string.IsNullOrWhiteSpace(operation.DesiredState))
                 {
-                    throw new NullReferenceException($"Desired state empty on operation {operation.Id}");
+                    throw new NullReferenceException($"Desired state empty on operation {operation.Id}: {operation.Description}");
                 }
 
                 var rulesFromOperationState = CloudResourceConfigStringSerializer.DesiredCorsRules(operation.DesiredState);
